Fix indexing bug when forming target from model output

+++ deep_learning/dataset.py
@@ -21,6 +21,6 @@ class HashReversalDataset(Dataset):
         return len(self.bits)
 
     def __getitem__(self, idx):
-        bits = torch.from_numpy(self.bits[idx].astype(float))
-        target = torch.from_numpy(self.target[idx].astype(float))
+        bits = torch.from_numpy(self.bits[idx]).float()
+        target = torch.from_numpy(self.target[idx]).float()
         return bits, target
